requirements.txt
opencv-python
opencv-contrib-python
pyaml